+++ examples/Misc.cpp
@@ -12,7 +12,9 @@ int main()
     std::cout << "major " << LIPSUM_CPP_VERSION_MAJOR << " minor "
               << LIPSUM_CPP_VERSION_MINOR << " patch "
               << LIPSUM_CPP_VERSION_PATCH << " commit "
-              << LIPSUM_CPP_VERSION_COMMIT << '\n';
+              << LIPSUM_CPP_VERSION_COMMIT << " time "
+              << LIPSUM_CPP_VERSION_TIME << " date " << LIPSUM_CPP_VERSION_DATE
+              << '\n';
 
     // should be 4
     std::cout << lpsm::CountSentences("Yes this is a test. [Testing "
@@ -36,31 +38,5 @@ int main()
                                   "dolor sit amet, aaaaaaaaaaaaaaaaaa\n\n## "
                                   "This is a subtitle")
               << "\n\n";
-    std::cout << "this is a plain URL: " << lpsm::GeneratePlainURL() << '\n';
-
-    // generate a URL with 2-5 words at the end (default)
-    std::cout << lpsm::GenerateURL() << '\n';
-
-    // generate a URL with 6-9 words at the end
-    std::cout << lpsm::GenerateURL(lpsm::ArgVec2(6, 9)) << '\n';
-
-    // equivalent statement
-    std::cout << lpsm::GenerateURLX(6, 9) << "\n\n";
-
-    std::cout << "your generated password is "
-              << lpsm::GenerateScramble(24, ' ', '~') << '\n';
-
-    // prints default "Lorem ipsum..." sentence
-    std::cout << lpsm::GenerateDefaultLipsumSentence() << "\n\n";
-
-    // generate slug of 2-5 words, separated by hyphen (default)
-    std::cout << lpsm::GenerateSlug() << '\n';
-
-    // generate slug of 6-9 words, separated by underscore
-    std::cout << lpsm::GenerateSlug(lpsm::ArgVec2(6, 9), '_') << '\n';
-
-    // equivalent statement
-    std::cout << lpsm::GenerateSlugX(6, 9, '_') << '\n';
-
     return 0;
 }