+++ src/lipsum.hpp
@@ -14,15 +14,6 @@
  * @example Advanced.cpp
  * @example Generator.cpp
  * @example GeneratorFormats.cpp
- * @example Formats.cpp
- * @example GenerateParagraph.cpp
- * @example GenerateParagraphs.cpp
- * @example GenerateSentence.cpp
- * @example GenerateSentenceFragment.cpp
- * @example GenerateSentences.cpp
- * @example GenerateText.cpp
- * @example GenerateWord.cpp
- * @example GenerateWords.cpp
  * @example Markdown.cpp
  * @example Misc.cpp
  * @example Source.cpp
@@ -35,12 +26,7 @@
 
 #    include "lipsum/core/internal.hpp"
 #    include "lipsum/core/stdincludes.hpp"
-#    include "lipsum/core_funcs.hpp"
-#    include "lipsum/core_xfuncs.hpp"
-#    include "lipsum/formats.hpp"
 #    include "lipsum/generator.hpp"
-#    include "lipsum/markdown.hpp"
-#    include "lipsum/md_xfuncs.hpp"
 #    include "lipsum/misc.hpp"
 
 /**
@@ -61,12 +47,8 @@ namespace lpsm = lipsum; ///< Alias for namespace lipsum
 #        include "lipsum/core/argvec2.inl"
 #        include "lipsum/core/internal.inl"
 #        include "lipsum/core/source.inl"
-#        include "lipsum/core_funcs.inl"
-#        include "lipsum/formats.inl"
 #        include "lipsum/generator.inl"
-#        include "lipsum/markdown.inl"
 #        include "lipsum/misc.inl"
-#        include "lipsum/xfuncs.inl"
 
 #    endif
 #endif